requirements.txt
line-bot-sdk
GoogleNews
openai
